fix: unblock CI after rebase - clippy + structure-check debt in candle-binding
Rebasing onto main (which now includes #2762's clippy cleanup) surfaced
two more pre-existing/self-inflicted CI gate failures in the same class
of files:

- bert_lora_test.rs had a pre-existing clippy::manual_range_contains
  finding (assert!(x >= 2 && x <= 10)) that #2762's cleanup didn't
  cover, since it only touched the three non-test files. Fixed to
  `(2..=10).contains(&num_classes)`.

- Adding classify_text_with_probabilities to bert_lora.rs grew it past
  its own structure-check ratchet baseline (main's post-#2762 863
  lines): the new method plus its thin classify_text wrapper add up to
  real new responsibility in a file that's already over the 800-line
  limit. Rather than growing it further or loosening the ratchet,
  moved classify_text/classify_text_with_probabilities/
  classify_text_internal into a new sibling module
  (bert_lora/probabilities.rs) as a second `impl
  HighPerformanceBertClassifier` block - a cohesive "single-text
  classification, with or without probabilities" unit, and back under
  bert_lora.rs's baseline with room to spare (822 lines).

No behavior change: cargo test --no-default-features --lib passes the
same 509/578 non-network-dependent tests (69 pre-existing
network/GPU-gated failures, none in the touched modules), including
the existing classify_text_with_probabilities coverage. Verified with
the CI-exact `rustup run 1.90 cargo clippy ... -D warnings` and
`structure_check.py --base-ref origin/main`: zero findings across all
seven Rust files this PR touches.
